Restart containers more aggressively; move restart logic to utils.py for shared usage
