+++ src/cli/minitrino/utils.py
@@ -14,6 +14,8 @@ from click import echo, style, prompt
 from textwrap import fill
 from shutil import get_terminal_size
 from functools import wraps
+from docker.errors import NotFound
+from concurrent.futures import ThreadPoolExecutor, as_completed
 
 
 class Logger:
@@ -386,3 +388,39 @@ def validate_yes(response=""):
     if response.lower() == "y" or response.lower() == "yes":
         return True
     return False
+
+
+def restart_containers(ctx, c_restart=[]):
+    """Restarts all the containers in the list."""
+
+    if c_restart == []:
+        return
+
+    c_restart = list(set(c_restart))
+
+    def restart_container(container_name):
+        """Helper function to restart a single container."""
+        try:
+            container = ctx.docker_client.containers.get(container_name)
+            ctx.logger.verbose(f"Restarting container '{container.name}'...")
+            container.restart()
+            ctx.logger.verbose(f"Container '{container.name}' restarted successfully.")
+        except NotFound:
+            raise err.MinitrinoError(
+                f"Attempting to restart container '{container_name}', but the container was not found."
+            )
+
+    with ThreadPoolExecutor() as executor:
+        futures = {
+            executor.submit(restart_container, container): container
+            for container in c_restart
+        }
+
+        for future in as_completed(futures):
+            container_name = futures[future]
+            try:
+                future.result()
+            except err.MinitrinoError as e:
+                ctx.logger.error(
+                    f"Error while restarting container '{container_name}': {str(e)}"
+                )
